Web_Tasks/Task_2/backend_django/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.4
node==1.2.1
odict==1.9.0
plumber==1.7
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==5.0
zope.proxy==5.0.0
